+++ beetsplug/hook.py
@@ -8,7 +8,6 @@ import string
 import subprocess
 from typing import TYPE_CHECKING, Any
 
-from beets.events import ALL_EVENTS
 from beets.plugins import BeetsPlugin
 
 if TYPE_CHECKING:
@@ -44,12 +43,14 @@ class HookPlugin(BeetsPlugin):
         for hook_index in range(len(hooks)):
             hook = self.config["hooks"][hook_index]
 
-            hook_event: EventType = hook["event"].as_choice(choices=ALL_EVENTS)
+            hook_event: EventType | str = hook["event"].as_str()
             hook_command = hook["command"].as_str()
 
             self.create_and_register_hook(hook_event, hook_command)
 
-    def create_and_register_hook(self, event: EventType, command: str) -> None:
+    def create_and_register_hook(
+        self, event: EventType | str, command: str
+    ) -> None:
         def hook_function(**kwargs) -> None:
             if command is None or len(command) == 0:
                 self._log.error('invalid command "{}"', command)